refactor(jobs): extract JobReaperService from LeaseManagerService (max-params 2/3)
LeaseManagerService injected 4 deps (config, claim, lease, apiKeys). Move
the zombie-reap engine — per-tenant reapZombies fan-out plus its tunables
(maxAttempts, backoffBaseMs) — into JobReaperService (config, claim,
apiKeys). LeaseManagerService keeps cron scheduling, leader election, the
re-entrancy guards, and the lease janitor (config, reaper, lease), 3 deps.
reapTick now delegates to reaper.reap() once it confirms leadership.

Part of the max-params=3 god-class split program.
